[REFACTOR] 기존 Guest 관련 클래스 패키지 수정 (#153)
